requirements.txt
# Web框架和服务器
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP客户端
httpx>=0.25.0

# 数据验证和序列化
pydantic>=2.5.0

# 环境变量管理
python-dotenv>=1.0.0

# 日志和监控
structlog>=23.2.0

# 性能优化
aiofiles>=23.2.1

# 类型提示支持
typing-extensions>=4.8.0

# 异步支持
asyncio-mqtt>=0.13.0

# JSON处理
orjson>=3.9.0

# 时间处理
python-dateutil>=2.8.2

# 中文分词（智能引用匹配需要）
jieba>=0.42.1

# 开发和测试工具（可选）
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.26.0

# 代码质量工具（可选）
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# 性能监控（可选）
prometheus-client>=0.18.0 